# ENH, BENCH: isclose runs faster by using errstate only when needed