<!--         ----------------------------------------------------------------
                MAKE SURE YOUR PR GETS THE ATTENTION IT DESERVES!
                ----------------------------------------------------------------

*  FORMAT IT RIGHT:
      https://www.numpy.org/devdocs/dev/development_workflow.html#writing-the-commit-message

*  IF IT'S A NEW FEATURE OR API CHANGE, TEST THE WATERS:
      https://www.numpy.org/devdocs/dev/development_workflow.html#get-the-mailing-list-s-opinion

*  HIT ALL THE GUIDELINES:
      https://numpy.org/devdocs/dev/index.html#guidelines

*  WHAT TO DO IF WE HAVEN'T GOTTEN BACK TO YOU:
      https://www.numpy.org/devdocs/dev/development_workflow.html#getting-your-pr-reviewed
-->

Following the discussions on the issue  #16160 , I have made the following change - 

-  Replace `np.all(xfin)` and `np.all(yfin)` with `np.all(xfin & yfin)`
-  Conditionally use the `errstate(invalid='ignore')` only when needed

I have added a small benchmark for isClose, which show significant improvement for small arrays, and no statistically significant change for large arrays. 

**Before**

             ========= ============
                param1              
              --------- ------------
                  1      15.4±0.7μs 
                 1000    17.3±0.4μs 
               1000000   7.38±0.4ms 
              ========= ============

**With Change**

             ========= ============
               param1              
             --------- ------------
                 1      10.6±0.1μs 
                1000    13.0±0.2μs 
              1000000   7.10±0.3ms 
             ========= ============