@seberg 2024-04-10
@beingkartik thanks for working on it, and it looks like good work.  Unfortunately, you came late.

It seems you based your work off the `maintenance/1.26.x` branch, and I expect this makes sense there (although only if atol/rtol are scalars or small arrays).

However, NumPy 2.0 allowed us to clean up the `np.errstate` mechanism a bit, and it is now *much* faster, so I would be very surprised if this doesn't turn out to be slower on NumPy 2.0.  (I haven't tried myself, I only verified that your branch is build on top of the `1.26.x` branch.)