@beingkartik 2024-04-10
Thanks for the quick review!

> It seems you based your work off the `maintenance/1.26.x` branch, and I expect this makes sense there (although only if atol/rtol are scalars or small arrays).

Yes, you're right, I did work off of the `maintenance/1.26.x`. 
> 
> However, NumPy 2.0 allowed us to clean up the `np.errstate` mechanism a bit, and it is now _much_ faster, so I would be very surprised if this doesn't turn out to be slower on NumPy 2.0. (I haven't tried myself, I only verified that your branch is build on top of the `1.26.x` branch.)

Your intuition is on point. Things are much faster on Numpy 2.0, at par with the performance with the change I made :smile:  . I will close this pull request.

LMK if there is value in adding the benchmark to main, I can make a separate pull request for that.