@seberg 2024-04-11
Dunno, we could add it, but then I am not sure adding `isclose` is particularly interesting without the code change.