@ngoldbaum 2024-04-11
> I will close this pull request.

Closing, feel free to open a followup with a new benchmark.